% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.coef}
\alias{fit.coef}
\title{fit.coef}
\usage{
fit.coef(X, y, intercept = T)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}

\item{intercept}{logical, "TRUE" by default.
If "FALSE", the model will not include the intercept.}
}
\value{
the estimated coefficients of linear model constructing by X and y
}
\description{
Gets the estimated coefficients of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
y <- rnorm(12)
fit.coef(X,y)

}
