% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.confint}
\alias{fit.confint}
\title{fit.confint}
\usage{
fit.confint(X, y, intercept = T, level = 0.95)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}

\item{intercept}{logical, "TRUE" by default.
If "FALSE", the model will not include the intercept.}

\item{level}{numerical value, stands for the significance of the confidence interval}
}
\value{
The upper and lower bounds of confidence interval for linear model constructing by X and y
}
\description{
Construct the confidence intervals parameters of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
y <- rnorm(12)
fit.confint(X,y)

}
