% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.hat.matrix}
\alias{fit.hat.matrix}
\title{fit.hat.matrix}
\usage{
fit.hat.matrix(X, intercept = T)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{intercept}{logical, "TRUE" by default.
If "FALSE", the model will not include the intercept.}
}
\value{
The upper and lower bounds of confidence interval for linear model constructing by X and y
}
\description{
Calculate the hat matrix of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
y <- rnorm(12)
fit.hat.matrix(X)

}
