% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.overall.test}
\alias{fit.overall.test}
\title{fit.overall.test}
\usage{
fit.overall.test(X, y, intercept = T)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}

\item{intercept}{logical, "TRUE" by default.
If "FALSE", the model will not include the intercept.}
}
\value{
The test statistic and corresponding p-value of the overall test for linear model constructing by X and y
}
\description{
Conduct the test of significance (overall test) of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
y <- rnorm(12)
fit.overall.test(X,y)

}
