% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.partial.test}
\alias{fit.partial.test}
\title{fit.partial.test}
\usage{
fit.partial.test(X, y, idx)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}

\item{idx}{a vector of indices of tested parameters(cannot includes 1, which stands for intercept)}
}
\value{
The test statistic and corresponding p-value of the partial test for linear model constructing by X and y
}
\description{
Conduct the partial test for several given parameters of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
y <- rnorm(12)
fit.overall.test(X,y)

}
