% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.predict}
\alias{fit.predict}
\title{fit.predict}
\usage{
fit.predict(X, y, Xnew)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}

\item{Xnew}{covariate matrix of new data that needs to predict}
}
\value{
the predicted values of Xnew by linear model constructing by X and y
}
\description{
Gets the predict values of a linear model given some new data
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
Xnew <- matrix(rnorm(100),nrow=10,ncol=10)
Xnew <- cbind(rep(1,10),Xnew)
y <- rnorm(12)
fit.predict(X,y,Xnew)

}
