% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.R.squared}
\alias{fit.R.squared}
\title{fit.R.squared}
\usage{
fit.R.squared(X, y, intercept = T)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}

\item{intercept}{logical, "TRUE" by default.
If "FALSE", the model will not include the intercept.}
}
\value{
The estimated R squared of linear model constructing by X and y
}
\description{
Gets the estimated R squared of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
y <- rnorm(12)
fit.R.squared(X,y)

}
