% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.sd}
\alias{fit.sd}
\title{fit.sd}
\usage{
fit.sd(X, y)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}
}
\value{
The estimated standard deviation of each parameter of linear model constructing by X and y
}
\description{
Gets the estimated standard deviation of each parameter(beta) of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
y <- rnorm(12)
fit.sd(X,y)

}
