% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.t.test}
\alias{fit.t.test}
\title{fit.t.test}
\usage{
fit.t.test(X, y)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}
}
\value{
The test statistic and corresponding p-value of each parameter of linear model constructing by X and y
}
\description{
Conducting t test for each parameter(beta) of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
y <- rnorm(12)
fit.t.test(X,y)

}
