% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.values}
\alias{fit.values}
\title{fit.values}
\usage{
fit.values(X, y)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}
}
\value{
the fitted values of linear model constructing by X and y
}
\description{
Gets the fitted values of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
y <- rnorm(12)
fit.values(X,y)

}
