% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fit.GLH.test}
\alias{fit.GLH.test}
\title{fit.GLH.test}
\usage{
fit.GLH.test(X, y, intercept = T, T.matrix, c = NULL)
}
\arguments{
\item{X}{input value of covariate matrix/vector}

\item{y}{input value of responses}

\item{intercept}{logical, "TRUE" by default.
If "FALSE", the model will not include the intercept.}

\item{T.matrix}{hypothesis matrix (or vector) giving linear combinations of coefficients by rows}

\item{c}{right-hand-side vector for hypothesis, with as many entries as rows in the hypothesis matrix;
can be omitted, in which case it defaults to a vector of zeroes}
}
\value{
The test statistic and corresponding p-value of the general linear hypothesis test for linear model constructing by X and y
}
\description{
Conduct the general linear hypothesis test for several given parameters of a linear model
}
\examples{
X <- matrix(rnorm(120),nrow=12,ncol=10)
X <- cbind(rep(1,12),X)
y <- rnorm(12)
T.matrix <- matrix(0,nrow=2,ncol=11)
T.matrix[1,4]<-1
T.matrix[1,5]<--1
T.matrix[2,5]<-1
T.matrix[2,6]<--1
fit.GLH.test(X,y,T.matrix=T.matrix)

}
